refactor: follow capability names
